+++ RELEASE-NOTES.md
@@ -11,6 +11,7 @@
 * #615 ECharts: basic support for geo charts
 * #617 ECharts: "map" series
 * #619 CsvLoader: own CSV parser implementation
+* #621 ECharts: VisualMap.splitNumber
 
 ## Release 2.0.0-M6